+++ tests/TestCase.php
@@ -1,37 +1,31 @@
 <?php
 
-namespace VendorName\Skeleton\Tests;
+namespace Iotron\LaravelRazorpay\Tests;
 
-use Illuminate\Database\Eloquent\Factories\Factory;
 use Orchestra\Testbench\TestCase as Orchestra;
-use VendorName\Skeleton\SkeletonServiceProvider;
+use Iotron\LaravelRazorpay\RazorpayServiceProvider;
 
 class TestCase extends Orchestra
 {
     protected function setUp(): void
     {
         parent::setUp();
-
-        Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'VendorName\\Skeleton\\Database\\Factories\\'.class_basename($modelName).'Factory'
-        );
     }
 
     protected function getPackageProviders($app)
     {
         return [
-            SkeletonServiceProvider::class,
+            RazorpayServiceProvider::class,
         ];
     }
 
     public function getEnvironmentSetUp($app)
     {
         config()->set('database.default', 'testing');
-
-        /*
-         foreach (\Illuminate\Support\Facades\File::allFiles(__DIR__ . '/database/migrations') as $migration) {
-            (include $migration->getRealPath())->up();
-         }
-         */
+        
+        // Set test config
+        config()->set('razorpay.key', 'test_key');
+        config()->set('razorpay.secret', 'test_secret');
+        config()->set('razorpay.webhook_secret', 'test_webhook_secret');
     }
 }